// app/components/GlobalStyles/enums.ts
export enum ButtonVariant {
    BLUE = 0,
    OUTLINE_BLUE = 1,
    OUTLINE_BLUE_TEXT = 2,
    GREEN = 3,
    OUTLINE_GREEN = 4,
    OUTLINE_GREEN_TEXT = 5,
  }
  
  export enum CssVars {
    WHITE = "--white",
    BLACK = "--black",
    TRANSPARENT = "--transparent",
    DEFAULT = "--blue-500-default",
    DEFAULT_DARK = "--blue-500-default-dark",
    BLUE_100 = "--blue-100",
    BLUE_200 = "--blue-200",
    BLUE_400 = "--blue-400",
    BLUE_900 = "--blue-900",
    GREEN_100 = "--green-100",
    GREEN_200 = "--green-200",
    GREEN_400 = "--green-400",
    GREEN_600 = "--green-600",
    GREEN_900 = "--green-900",
  }
  